+++ src/onegov.swissvotes/onegov/swissvotes/tests/test_forms.py
@@ -67,7 +67,6 @@ def test_attachments_form(swissvotes_app, attachments):
         id=1,
         bfs_number=Decimal('1'),
         date=date(1990, 6, 2),
-        decade=NumericRange(1990, 1999),
         legislation_number=4,
         legislation_decade=NumericRange(1990, 1994),
         title_de="Vote DE",
@@ -315,7 +314,6 @@ def test_search_form(swissvotes_app):
     votes = SwissVoteCollection(swissvotes_app)
     kwargs = {
         'date': date(1990, 6, 2),
-        'decade': NumericRange(1990, 1999),
         'legislation_number': 4,
         'legislation_decade': NumericRange(1990, 1994),
         'title_de': "Vote DE",
@@ -446,7 +444,6 @@ def test_update_dataset_form(session):
         '1.2.2008',  # datum / DATE
         1,  # legislatur / INTEGER
         '2004-2008',  # legisjahr / INT4RANGE
-        '2000-2009',  # jahrzent / INT4RANGE
         'kurztitel de',  # titel_kurz_d
         'kurztitel fr',  # titel_kurz_f
         'titel de',  # titel_off_d